Test 'Countries', 'People', 'Role' and 'RoleEntityPermissions'

+++ SMS.Infrastructure/Repositories/ApplicationLogRepository.cs
@@ -87,7 +87,7 @@ namespace SMS.Infrastructure.Repositories
         private static ApplicationLog MapApplicationLog(SqlDataReader reader)
         {
             var auditLogIdOrdinal = reader.GetOrdinal("AuditLogId");
-            int? auditLogId = reader.IsDBNull(auditLogIdOrdinal) ? null : reader.GetInt32(auditLogIdOrdinal);
+            long? auditLogId = reader.IsDBNull(auditLogIdOrdinal) ? null : reader.GetInt64(auditLogIdOrdinal);
 
             var exceptionOrdinal = reader.GetOrdinal("Exception");
             string? exceptionMessage = reader.IsDBNull(exceptionOrdinal) ? null : reader.GetString(exceptionOrdinal);
